+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 
 ## [UNRELEASED]
 
+### Added
+
+- Added a new command `/changelog` to show HexBug's changelog.
+
 ### Mods Added
 
 - Hex Trace: 0.1.0+1.20.1